Extract object CLI commands
